requirements.txt
serpapi==0.1.5
python-dotenv==1.0.0
streamlit==1.30.0
langchain==0.1.1
langchain_core==0.1.13
langchain_community==0.0.13
openai==1.9.0
langchain_openai==0.0.3